Change tensor conversion to float for embeddings

+++ Llama/Llama-3.2-1B/wake2vec_Phase1_on_llama_3_2_1b.py
@@ -544,7 +544,7 @@ from numpy.linalg import norm as l2
 from scipy import stats
 from sklearn.decomposition import PCA
 
-E_post = final_emb.numpy()
+E_post = final_emb.float().numpy()
 vocab_size, dim = E_post.shape
 num_new_tokens = vocab_size - BASE_VOCAB
 
@@ -555,7 +555,7 @@ E_new = E_post[BASE_VOCAB:]
 pre_path = RUN_DIR / "embeddings_pre.pt"
 has_pre = pre_path.exists()
 if has_pre:
-    E_pre_all = torch.load(pre_path, map_location="cpu").numpy()
+    E_pre_all = torch.load(pre_path, map_location="cpu").float().numpy()
     E_pre_base = E_pre_all[:BASE_VOCAB]
     E_pre_new = E_pre_all[BASE_VOCAB:]
     print(f"[PRE] Loaded pre-training embeddings: {E_pre_all.shape}")
